Added the same code with additional comments and a drawing of the quantum circuit
Jupyter notebook with the same code (unchanged). The content of the code (functionality) is the same as the file task1_Sanaa.ipynb and can be checked by running the two notebooks separately. The readability is improved in this version by adding comments and a drawing of the quantum circuit that is being executed.